Fixes issues with correct Field Selector state on upgrade
When upgrading, user reported the HTML fields were disabled even when correct setting was enabled. During testing it was found the "change" event for the radio button was running twice, and because we now disabled the HTML field by default when the change event ran it caused this bug.

Resolution for v5 is to prevent change event running if current element isn't selected or is undefined.

Another issue surfaced while investigating. That was early Core Booster adopters who upgraded would have the Field Selector feature disabled. This has already been patched.